Corrected definition in string for CU and CD.  The syntax for alias was being misused.
"Concave down"=>{alias=>"CU"}  means that "Concave down" is an alias for "CU", which must
have been previously defined as CU=>{}.  The problem had the concept reversed.

I made further changes so that now "Increasing", "Decreasing", "Concave up" and so forth 
are the primary strings and CU, CD, INC are aliased to these.  This way the correct 
answers appear as "increasing" which is less cryptic than "INC".

Don't know if this is the best strategy or not.